The proxy mode allows Spring container to inject a new object into the singleton bean
The prototype bean is not autowired into the singleton bean at the time of its creation. Instead a proxy(like a placeholder) is autowired. It adds a level of indirection. When requesting the prototype bean from spring, a new instance of prototype bean is created and returned by the application context.